+++ IP_LOOTBOX/IPLootbox/BoxSize.cs
@@ -1,0 +1,7 @@
+﻿namespace IPLootbox
+{
+    public enum BoxSize
+    {
+        Small = 1, Middle, Big
+    }
+}
