+++ IP_LOOTBOX/IPLootbox/IPLootbox.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31205.134
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IPLootbox", "IPLootbox.csproj", "{DDD99319-B373-4091-9F70-EBC654BE0460}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DDD99319-B373-4091-9F70-EBC654BE0460}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DDD99319-B373-4091-9F70-EBC654BE0460}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DDD99319-B373-4091-9F70-EBC654BE0460}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DDD99319-B373-4091-9F70-EBC654BE0460}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CF56159E-583D-4E01-89DA-BB8A9C578FD6}
+	EndGlobalSection
+EndGlobal
