+++ IP_LOOTBOX/IPLootbox/MenuMain.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace IPLootbox
+{
+    public static class MenuMain
+    {
+        public static void PrintMenu()
+        {
+            Console.Clear();
+            Console.WriteLine("Добро пожаловать.");
+            Console.WriteLine("ИП ЛУТБОКС. Доставка бандеролей.\n\n\n");
+            Console.WriteLine("ВЫБЕРЕТЕ РОЛЬ:\n");
+
+            Console.WriteLine("\t1. ПОЛЬЗОВАТЕЛЬ");
+            Console.WriteLine("\t2. АДМИНИСТРАТОР");
+
+            Console.WriteLine("\n\n\nНажмите ESC для выхода их программы.");
+        }
+    }
+}
