+++ IP_LOOTBOX/LOOTBOX/LOOTBOX.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31205.134
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LOOTBOX", "LOOTBOX.csproj", "{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IPLootbox", "..\IPLootbox\IPLootbox.csproj", "{605A9749-B5C9-403A-9E3C-6604D933E025}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{605A9749-B5C9-403A-9E3C-6604D933E025}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{605A9749-B5C9-403A-9E3C-6604D933E025}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{605A9749-B5C9-403A-9E3C-6604D933E025}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{605A9749-B5C9-403A-9E3C-6604D933E025}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D9A16DBD-88D0-41BB-933B-FEF049F9B4C0}
+	EndGlobalSection
+EndGlobal
