--- LOOTBOX/BoxSize.cs
@@ -1,7 +1,0 @@
-﻿namespace LOOTBOX
-{
-    public enum BoxSize
-    {
-        Small = 1, Middle, Big
-    }
-}
