--- LOOTBOX/LOOTBOX.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LOOTBOX", "LOOTBOX.csproj", "{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B75FED8B-92D7-4F13-B4DB-7C8CFEF22D0C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D9A16DBD-88D0-41BB-933B-FEF049F9B4C0}
-	EndGlobalSection
-EndGlobal
